requirements.txt
asgiref==3.8.1
asttokens==3.0.0
colorama==0.4.6
decorator==5.2.1
Django==5.2.3
django-extensions==4.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
executing==2.2.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
prompt_toolkit==3.0.51
pure_eval==0.2.3
Pygments==2.19.2
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
psycopg2-binary==2.9.10
Pillow==11.3.0
openpyxl==3.1.5
pytz==2025.2